+++ src/dashboard-refactor/colors.tsx
@@ -1,6 +1,7 @@
 const colors = {
     onHover: '#BABABA',
-    highlightGrey: '#C4C4C4',
+    midGrey: '#C4C4C4',
+    lightGrey: '#E1E1E1',
     iconDefault: '#3A2F45',
 }
