+++ src/dashboard-refactor/header/SidebarHeader/SidebarHeader.tsx
@@ -1,7 +1,26 @@
 import React, { PureComponent } from 'react'
+import styled from 'styled-components'
+
+import styleConstants from '../../styleConstants'
+
 import Margin from 'src/dashboard-refactor/components/Margin'
 import SidebarToggle from './SidebarToggle/SidebarToggle'
-import { CollectionTitle, SidebarHeaderContainer } from './styledComponents'
+
+const { fonts } = styleConstants
+
+export const SidebarHeaderContainer = styled.div`
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+`
+
+export const CollectionTitle = styled.p`
+    margin: 0;
+    font-family: ${fonts.primary.name};
+    font-weight: ${fonts.primary.bold};
+    line-height: 21px;
+`
 
 import {
     SidebarLockedState,
