+++ src/dashboard-refactor/styleConstants.tsx
@@ -1,10 +1,12 @@
-const styleConstants = {
-    fonts: {
-        primary: {
-            name: 'Poppins',
-            bold: 700,
-        },
+export const fonts = {
+    primary: {
+        name: 'Poppins',
+        bold: 700,
     },
+}
+
+const styleConstants = {
+    fonts,
 }
 
 export default styleConstants
