+++ src/dev/stories/index.tsx
@@ -19,3 +19,4 @@ import './copy-paster'
 import './sharing'
 import './dashboard-refactor/header/sidebar-header'
 import './dashboard-refactor/header/sidebar-toggle'
+import './dashboard-refactor/header/search-bar'
